discord-bot/requirements.txt
discord-py-interactions
tortoise-orm[asyncpg]
httpx
